+++ crates/stackable-shared/src/lib.rs
@@ -1,5 +1,7 @@
 //! This crate contains various shared helpers and utilities used across other crates in this
 //! workspace.
 
+pub mod commons;
 pub mod crd;
+pub mod time;
 pub mod yaml;