+++ crates/stackable-shared/src/time/mod.rs
@@ -1,7 +1,6 @@
 mod duration;
 mod serde_impl;
 
-#[cfg(feature = "time")]
 mod time_impl;
 
 pub use duration::*;